Rename BrowserWindow to EntityBrowser
Changes:
- Rename tools package from 'browser' to 'entitybrowser'
- Update references in ToolRegistry and ToolToolBar
- Rename BrowserWindow class to EntityBrowser

Ref #1835
